tokenizer/requirements.txt
transformers
google-genai
colorama
